Reorganized remote packages
