+++ src/main/java/bullet_hell/remote/both/Task.java
@@ -1,4 +1,4 @@
-package bullet_hell.model.remote;
+package bullet_hell.remote.both;
 
 public interface Task<T> {
     T execute();
